Added action_rules field to ActionMessage; refactored dataclasses into various smaller files.
