+++ hasher-matcher-actioner/hmalib/common/actioner_models.py
@@ -5,127 +5,12 @@ import json
 import typing as t
 
 from dataclasses import dataclass, field, fields
+from hmalib.common.message_models import BankedSignal
 from hmalib.common.logging import get_logger
-from hmalib.models import BankedSignal, MatchMessage
+from hmalib.common.message_models import MatchMessage
 from requests import get, post, put, delete, Response
 
 logger = get_logger(__name__)
-
-
-@dataclass(unsafe_hash=True)
-class Label:
-    key: str
-    value: str
-
-    def __eq__(self, another_label: object) -> bool:
-        if not isinstance(another_label, Label):
-            return NotImplemented
-        return self.key == another_label.key and self.value == another_label.value
-
-
-@dataclass(unsafe_hash=True)
-class ClassificationLabel(Label):
-    key: str = field(default="Classification", init=False)
-
-
-@dataclass(unsafe_hash=True)
-class BankSourceClassificationLabel(Label):
-    key: str = field(default="BankSourceClassification", init=False)
-
-
-@dataclass(unsafe_hash=True)
-class BankIDClassificationLabel(Label):
-    key: str = field(default="BankIDClassification", init=False)
-
-
-@dataclass(unsafe_hash=True)
-class BankedContentIDClassificationLabel(Label):
-    key: str = field(default="BankedContentIDClassification", init=False)
-
-
-@dataclass(unsafe_hash=True)
-class ActionLabel(Label):
-    key: str = field(default="Action", init=False)
-
-
-@dataclass(unsafe_hash=True)
-class ThreatExchangeReactionLabel(Label):
-    key: str = field(default="ThreatExchangeReaction", init=False)
-
-
-@dataclass
-class Action:
-    action_label: ActionLabel
-    priority: int
-    superseded_by: t.List[ActionLabel]
-
-
-@dataclass
-class ActionRule(config.HMAConfig):
-    """
-    Action rules are config-backed objects that have a set of labels (both
-    "must have" and "must not have") which, when evaluated against the
-    classifications of a matching banked piece of content, lead to an action
-    to take (specified by the rule's action label). By convention each action
-    rule's name field is also the value field of the rule's action label.
-    """
-
-    action_label: ActionLabel
-    must_have_labels: t.Set[Label]
-    must_not_have_labels: t.Set[Label]
-
-
-TUrl = t.Union[t.Text, bytes]
-
-
-@dataclass
-class ActionMessage(MatchMessage):
-    """
-    The action performer needs the match message plus which action to perform
-    """
-
-    action_label: ActionLabel = ActionLabel("UnspecifiedAction")
-    action_rules: t.List[ActionRule] = field(default_factory=list)
-
-    @classmethod
-    def from_match_message_action_label_and_action_rules(
-        cls,
-        match_message: MatchMessage,
-        action_label: ActionLabel,
-        action_rules: t.List[ActionRule],
-    ) -> "ActionMessage":
-        return cls(
-            match_message.content_key,
-            match_message.content_hash,
-            match_message.matching_banked_signals,
-            action_label,
-            action_rules,
-        )
-
-
-@dataclass
-class ReactionMessage(MatchMessage):
-    """
-    The reactioner needs the match message plus which reaction to send back
-    to the source of the signal (for now, ThreatExchange).
-    """
-
-    reaction_label: ThreatExchangeReactionLabel = field(
-        default=ThreatExchangeReactionLabel("UnspecifiedThreatExchangeReaction")
-    )
-
-    @classmethod
-    def from_match_message_and_label(
-        cls,
-        match_message: MatchMessage,
-        threat_exchange_reaction_label: ThreatExchangeReactionLabel,
-    ) -> "ReactionMessage":
-        return cls(
-            match_message.content_key,
-            match_message.content_hash,
-            match_message.matching_banked_signals,
-            threat_exchange_reaction_label,
-        )
 
 
 class ActionPerformer(config.HMAConfigWithSubtypes):
