+++ hasher-matcher-actioner/hmalib/common/classification_models.py
@@ -1,0 +1,34 @@
+# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
+
+from dataclasses import dataclass, field
+
+
+@dataclass(unsafe_hash=True)
+class Label:
+    key: str
+    value: str
+
+    def __eq__(self, another_label: object) -> bool:
+        if not isinstance(another_label, Label):
+            return NotImplemented
+        return self.key == another_label.key and self.value == another_label.value
+
+
+@dataclass(unsafe_hash=True)
+class ClassificationLabel(Label):
+    key: str = field(default="Classification", init=False)
+
+
+@dataclass(unsafe_hash=True)
+class BankSourceClassificationLabel(Label):
+    key: str = field(default="BankSourceClassification", init=False)
+
+
+@dataclass(unsafe_hash=True)
+class BankIDClassificationLabel(Label):
+    key: str = field(default="BankIDClassification", init=False)
+
+
+@dataclass(unsafe_hash=True)
+class BankedContentIDClassificationLabel(Label):
+    key: str = field(default="BankedContentIDClassification", init=False)
